fixing some warnings and stuff

+++ S-Tengine2/Source/S-Tengine2/src/poly2Tri/sweep/sweep.cc
@@ -801,13 +801,9 @@ void Sweep::FlipScanEdgeEvent(SweepContext& tcx, Point& ep, Point& eq, Triangle&
 }
 
 Sweep::~Sweep() {
-
     // Clean up memory
-    for(int i = 0; i < nodes_.size(); i++) {
+    for(unsigned long int i = 0; i < nodes_.size(); i++) {
         delete nodes_[i];
     }
-
 }
-
 }
-
